Refactor: move `ClientManagerTest` to correct namespace
