+++ carp.clients.core/src/commonTest/kotlin/dk/cachet/carp/clients/domain/study/StudyTest.kt
@@ -1,8 +1,8 @@
 package dk.cachet.carp.clients.domain.study
 
 import dk.cachet.carp.clients.application.study.StudyStatus
-import dk.cachet.carp.clients.domain.connectedDevice
-import dk.cachet.carp.clients.domain.smartphone
+import dk.cachet.carp.clients.connectedDevice
+import dk.cachet.carp.clients.smartphone
 import dk.cachet.carp.common.application.UUID
 import dk.cachet.carp.common.application.devices.AnyPrimaryDeviceConfiguration
 import dk.cachet.carp.common.application.users.AssignedTo
